refactor: restructure the src using setuptools
