+++ src/myoktros/gesture.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+import aenum
+
+
+class Gesture(aenum.Enum):
+    zero = 0
+    one = 1
+    two = 2
+    three = 3
+    four = 4
